requirements.txt
azure-common==1.1.25
azure-core==1.8.0
azure-cosmos==4.1.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
config==0.5.0.post0
cryptography==3.0
idna==2.10
pycparser==2.20
pypyodbc==1.3.4
pyratemp==0.3.2
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
